requirements.txt
asgiref==3.2.10
astroid==2.4.2
colorama==0.4.4
Django==3.1.2
django-widget-tweaks==1.4.8
djangorestframework==3.12.1
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.0.0
pylint==2.6.0
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
wrapt==1.12.1
